fix: 즐겨찾기, 체크 기능 완성

+++ back/src/main/java/msps/back/dto/response/AllMenuIngredientsGetResponse.java
@@ -37,7 +37,7 @@ public class AllMenuIngredientsGetResponse {
         private String name;
         private String type;
         private AmountType amountType;
-        private Long amountValue;
+        private int amountValue;
         private String amountUnit;
         private String amountText;
     }
